Revert a series of gtk-update-icon-cache changes for F20
This reverts a series of commits that changed gtk-update-icon-cache to
not include image data in the generated image cache. It is safer to keep
this out of F20 to avoid introducing non-bugfix changes that might cause
a fallout in other packages. In particular, the gtk3 version in F20 is
not tested to work with this.